Fix integration issues and clean the repo up.
Because we can't walk a directory "up" from our page in docs/ on
GitHub.io hosting I had to move the data into docs/. This
nebulosity is the result of not having index.html on the top level
of the homework repo template that was given to us.

Remove a bunch of old extraneous files.